mock: Fix build endpoints not handling missing packages properly
It's entirely valid for the "build side" of a package to be missing, in
particular, if the arch has been disabled, or the user just hasn't set
anything there yet. In these cases, unknown_package should *not* be
returned, instead using endpoint-specific handling for the different
scenarios.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>

update to missing arch thing

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>