Add defaults for RebuildMode and BlockMode
rebuild and block are optional attributes in the `<repository />`
element with defined default values. Reflect this in the deserialization
as well otherwise the runner can't cope with repositories that don't
have those attributes set in their meta.

For the mock this now leaves out those attributes if they have their
default values. OBS doesn't reserialize meta configuration though, so
meta's in the wild might still have the default values explicitly in
which we now no longer test, but it seems unlikely for that to be an
issue.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>