Avoid failing to deserialize broken links
It's possible for a branch link to be invalid, in which case one or more
of `{srcmd5, xsrcmd5, lsrcmd5}` may be missing, so this adds support
for capturing and makes the `*srcmd5` parameters optional. (As-is, on
link error, OBS seems to only set `xsrcmd5`, if possible, but simply
making them all optional is a bit less brittle than trying to encode all
this logic in enums.) This also adds some basic artificial error
injection in the mocks for testing this out.